\name{nROUSE-package}
\alias{nROUSE-package}
\alias{nROUSE}
\docType{package}
\title{
\packageTitle{nROUSE}
}
\description{
\packageDescription{nROUSE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nROUSE}
\packageIndices{nROUSE}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{nROUSE}

Maintainer: \packageMaintainer{nROUSE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
