% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_code.R
\name{nROUSE_demonstration}
\alias{nROUSE_demonstration}
\title{Demonstration of the nROUSE model.}
\usage{
nROUSE_demonstration(primeTypes = c("Target", "Foil", "Neither", "Both"),
  design = c(50, 450, 500), prm = NULL)
}
\arguments{
\item{primeTypes}{a character vector with up to 4 types of prime: 
'Target', 'Foil', 'Neither', or 'Both'.}

\item{design}{a vector with the duration (in ms) of the target flash,
the target mask, and the choice options.}

\item{prm}{an optional named vector for the parameters of the 
nROUSE model, where...
\itemize{
  \item Fe = the feedback scalar.
  \item N = The noise multiplier.
  \item L = The constant leak current.
  \item D = The synaptic depletion rate.
  \item R = The replenishment rate.
  \item I = The inhibition constant.
  \item Th = The noise multiplier.
  \item Ta = The temporal attention parameter.
  \item SV = The visual integration rate.
  \item SO = The orthographic integration rate.
  \item SS = The semantic integration rate.
}}
}
\description{
Generates a plot of the activation at the highest level of the 
neural network (i.e. semantic-lexical) for the target and foil 
representations under different prime types and an inputted 
prime duration.
}
\details{
The user will receive a prompt to input a prime duration. 
  The function will then generate a plot showing the time course 
  of activation for the target/foil representations for the 
  set of prime types. The predicted accuracies and the difference 
  distribution from which these accuracies are derived are shown 
  on the right.
}
\examples{
\dontrun{
# Default
nROUSE_demonstration()

# Fewer prime types
pt = c( 'Target', 'Foil' )
nROUSE_demonstration(primeTypes = pt)

# Different timing for target stimulus flash and subsequent mask
nROUSE_demonstration(design=c(84,416,500))

# Different parameter values (Must be named)
prm = c( I = .5, Ta = .8, N = .025 )
nROUSE_demonstration(prm=prm)
}

}
