% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_code.R
\name{nROUSE_predictions}
\alias{nROUSE_predictions}
\title{Predictions from the nROUSE model}
\usage{
nROUSE_predictions(primeDurations = c(17, 50, 150, 400, 2000),
  primeTypes = c("Target", "Foil", "Neither", "Both"), design = c(50, 450,
  500), opt = c(F, T), prm = NULL)
}
\arguments{
\item{primeDurations}{a vector of prime durations between 
17 and 2000 ms.}

\item{primeTypes}{a character vector with up to 4 types of prime: 
'Target', 'Foil', 'Neither', or 'Both'.}

\item{design}{a vector with the duration (in ms) of the target flash,
the target mask, and the choice options.}

\item{opt}{a logical vector indicating if the predicted accuracies 
latencies should be returned, and if a plot should be generated.}

\item{prm}{an optional named vector for the parameters of the 
nROUSE model.}
}
\value{
A list consisting of a matrix of predicted accuracies by 
  prime duration and type, and an array with the predicted perceptual 
  identification latencies for the target and foil by prime duration 
  and prime type.
}
\description{
Generates a plot of the predicted accuracy over different 
prime durations and types for the nROUSE model (output can 
be generated as well).
}

