% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_nROUSE}
\alias{simulate_nROUSE}
\title{The nROUSE model}
\usage{
simulate_nROUSE(presentations, primeInput, param = as.numeric(c(0.25, 0.0302,
  0.15, 0.324, 0.022, 0.9844, 0.15, 1, 0.0294, 0.0609, 0.015)))
}
\arguments{
\item{presentations}{A vector giving the duration (in ms) of 
the prime, the target, the mask, and the choice alternatives.}

\item{primeInput}{A vector of two elements used to set the type 
of prime. For instance, [2,0] indicates a double prime for 
targets, while [0,1] indicates a single prime for foils.}

\item{param}{A vector giving the values for the parameters.
\describe{
  \item{\code{param[1]}}{Semantic to orthographic feedback scalar}
  \item{\code{param[2]}}{Noise constant}
  \item{\code{param[3]}}{Constant leak current}
  \item{\code{param[4]}}{Synaptic depletion rate}
  \item{\code{param[5]}}{Recovery rate}
  \item{\code{param[6]}}{Inhibition constant}
  \item{\code{param[7]}}{Activation threshold}
  \item{\code{param[8]}}{Temporal attention}
  \item{\code{param[9]}}{Integration time constant (Visual)}
  \item{\code{param[10]}}{Integration time constant (Orthographic)}
  \item{\code{param[10]}}{Integration time constant (Semantic)}
   }}
}
\description{
Function to simulate the predictions of the nROUSE model (Huber 
and O'Reilly, 2003) for perceptual identification latencies 
and forced-choice accuracy.
}
\section{References}{

Huber, D. E., & O'Reilly, R. C. (2003). Persistence and 
  accommodation in short-term priming and other perceptual 
  paradigms: Temporal segregation through synaptic depression. 
  Cognitive Science, 27(3), 403-430.
}
\examples{
# Define duration (in ms) for prime, target, mask, and choices
presentations = c( 17, 50, 450, 500 )
# Set a double prime for targets
primeInput = c(2,0)
# Simulate model with default parameters
sim = simulate_nROUSE( presentations, primeInput )

}

