% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_code.R
\name{nROUSE_demonstration}
\alias{nROUSE_demonstration}
\title{Demonstration of the nROUSE model.}
\usage{
nROUSE_demonstration(primeTypes = c("Target", "Foil", "Neither", "Both"),
  design = c(50, 450, 500), prm = NULL)
}
\arguments{
\item{primeTypes}{a character vector with up to 4 types of prime: 
'Target', 'Foil', 'Neither', or 'Both'.}

\item{design}{a vector with the duration (in ms) of the target flash,
the target mask, and the choice options.}

\item{prm}{an optional named vector for the parameters of the 
nROUSE model.}
}
\description{
Generates a plot of the activation at the highest level of the 
neural network (i.e. semantic-lexical) for the target and foil 
representations under different prime types and an inputted 
prime duration.
}
\details{
The user will receive a prompt to input a prime duration. 
  The function will then generate a plot showing the time course 
  of activation for the target/foil representations for the 
  set of prime types. The predicted accuracies and the difference 
  distribution from which these accuracies are derived are shown 
  on the right.
}

